+++ src/main/java/be/company/fca/controller/PouleController.java
@@ -3,6 +3,7 @@ package be.company.fca.controller;
 import be.company.fca.model.Championnat;
 import be.company.fca.model.Division;
 import be.company.fca.model.Poule;
+import be.company.fca.repository.ChampionnatRepository;
 import be.company.fca.repository.DivisionRepository;
 import be.company.fca.repository.PouleRepository;
 import io.swagger.annotations.Api;
@@ -19,6 +20,12 @@ public class PouleController {
 
     @Autowired
     private PouleRepository pouleRepository;
+
+    @Autowired
+    private DivisionRepository divisionRepository;
+
+    @Autowired
+    private ChampionnatRepository championnatRepository;
 
     @RequestMapping(method= RequestMethod.GET, path="/public/poules")
     public Iterable<Poule> getPoulesByDivision(@RequestParam Long divisionId) {
@@ -39,21 +46,54 @@ public class PouleController {
     @PreAuthorize("hasAuthority('ADMIN_USER')")
     @RequestMapping(value = "/private/poule", method = RequestMethod.POST)
     public Poule addPoule(@RequestParam Long divisionId, @RequestBody Poule poule){
-        Division division = new Division();
-        division.setId(divisionId);
+
+        Division division = divisionRepository.findOne(divisionId);
         poule.setDivision(division);
-        return pouleRepository.save(poule);
+
+        // Operation non-permise si le calendrier est valide ou cloture
+        if (division.getChampionnat().isCalendrierValide() || division.getChampionnat().isCloture()){
+            throw new RuntimeException("Operation not supported");
+        }
+
+        Poule pouleSaved = pouleRepository.save(poule);
+
+        // On signale que le calendrier doit etre rafraichi si la poule a ete sauvee
+        championnatRepository.updateCalendrierARafraichir(division.getChampionnat().getId(),true);
+
+        return pouleSaved;
     }
 
     @PreAuthorize("hasAuthority('ADMIN_USER')")
     @RequestMapping(value = "/private/poule/allerRetour", method = RequestMethod.PUT)
     public void updateAllerRetour(@RequestParam Long pouleId, @RequestParam boolean allerRetour){
+
+        Poule poule = pouleRepository.findOne(pouleId);
+
+        // Operation non-permise si le calendrier est valide ou cloture
+        if (poule.getDivision().getChampionnat().isCalendrierValide() || poule.getDivision().getChampionnat().isCloture()){
+            throw new RuntimeException("Operation not supported");
+        }
+
         pouleRepository.updateAllerRetour(pouleId,allerRetour);
+
+
+        // On signale que le calendrier doit etre rafraichi si la poule a ete sauvee
+        championnatRepository.updateCalendrierARafraichir(poule.getDivision().getChampionnat().getId(),true);
+
     }
 
     @PreAuthorize("hasAuthority('ADMIN_USER')")
     @RequestMapping(value = "/private/poule", method = RequestMethod.DELETE)
     public void deletePoule(@RequestParam Long id){
+        Poule poule = pouleRepository.findOne(id);
+
+        // Operation non-permise si le calendrier est valide ou cloture
+        if (poule.getDivision().getChampionnat().isCalendrierValide() || poule.getDivision().getChampionnat().isCloture()){
+            throw new RuntimeException("Operation not supported");
+        }
+
         pouleRepository.delete(id);
+        // On signale que le calendrier doit etre rafraichi si la poule a ete supprimee
+        championnatRepository.updateCalendrierARafraichir(poule.getDivision().getChampionnat().getId(),true);
     }
 }
